backend/requirements.txt
django==5.0.2
djangorestframework==3.14.0
transformers==4.38.2
torch==2.2.1
pandas==2.2.1
python-dotenv==1.0.1
requests==2.31.0
django-redis==5.4.0
django-cors-headers==4.3.1 